backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
langgraph==0.2.60
langgraph-checkpoint-postgres==2.0.11
langchain==0.3.13
langchain-openai==0.2.14
langchain-core==0.3.28
redis==5.2.1
websockets==14.1
httpx==0.28.1
